Gjennomgang av readme og oppgavene
