+++ src/main/kotlin/no/bekk/introduction/DataClass.kt
@@ -1,6 +1,8 @@
 package no.bekk.introduction
 
-class Konsulent(val name: String)
+class Konsulent(
+    val name: String,
+)
 
 fun main() {
     val gaute = Konsulent("Gaute")
@@ -8,5 +10,4 @@ fun main() {
 
     val gaute2 = Konsulent("Gaute")
     println("gaute1 == gaute2: ${gaute == gaute2}")
-
 }
