+++ src/main/kotlin/no/bekk/introduction/ExtensionFunctions.kt
@@ -1,17 +1,16 @@
 package no.bekk.introduction
 
-
-// Oppgave 1 - skriv om denne til en extension function
+// Oppgave 1 - Skriv om denne til en extension function
 fun getInfo(coach: BootcampCoach): String {
     return "${coach.name} er i avdeling ${coach.avdeling} og har jobbet ${coach.yearsInBekk} år i Bekk"
 }
 
-// Oppgave 2 - skriv om denne til en extension function
+// Oppgave 2 - Skriv om denne til en extension function
 fun hasWorkedLongerThan(coach1: BootcampCoach, coach2: BootcampCoach): Boolean {
     return coach1.yearsInBekk > coach2.yearsInBekk
 }
 
-// Oppgave 3 - skriv om disse til extension functions, og bruk chaining
+// Oppgave 3 - Skriv om disse til extension functions, og bruk chaining
 fun addYears(coach: BootcampCoach, years: Int): BootcampCoach {
     return BootcampCoach(coach.name, coach.avdeling, coach.yearsInBekk + years)
 }
@@ -33,5 +32,5 @@ fun main() {
     // Oppgave 3 - kan du gjøre dette mer leselig ved hjelp av extension functions?
     print(getInfo(addYears(withAvdeling(ingrid, Avdeling.BMC), 1)))
 
-    // ferdig? Bra jobba! Se løsningsforslag i README.md og gå videre til neste 🚀
+    // Ferdig? Bra jobba! Se løsningsforslag i README.md og gå videre til neste 🚀
 }
