+++ src/main/kotlin/no/bekk/introduction/Funksjoner.kt
@@ -1,9 +1,6 @@
 package no.bekk.introduction
 
-class BekkAnsatt(val name: String) {
-
-}
-
+data class BekkAnsatt(val name: String)
 fun main() {
     val ingrid = BekkAnsatt("Ingrid")
     // ingrid.greet()
