+++ src/main/kotlin/no/bekk/introduction/HighOrderFunction.kt
@@ -1,13 +1,16 @@
 package no.bekk.introduction
 
-val newList = coacher2023.map {
-    when (it.name) {
-        "Morten" -> it.copy(avdeling = Avdeling.BMC)
-        "Ragnhild" -> it.copy(avdeling = Avdeling.DESIGN)
-        else -> it
-    }
-}
+// val realCoacher2023
 
-val teknologiCoacher = newList.filter {
-    it.avdeling == Avdeling.TEKNOLOGI
+// val teknologiCoacher
+
+// val numberOfYears
+fun main() {
+    println("Årets coacher er $coacher2023!")
+
+    // println("RealCoacher2023 er $realCoaher2023")
+
+    // println("TeknologiCoachene er $teknologiCoacher")
+
+    // println("Antall år i bekk er $numberOfYears ")
 }
